Changed the info for the logger.
Decided that it's not too slow. When I said it was slow, I was looking at generating overlap populations for 40 different fragments, which is a bit extreme as it's usually done between 5-10 fragments. There ought to be a Numeric trick to speed it up, but I haven't had any luck so I'm giving up for now.


git-svn-id: https://svn.code.sf.net/p/cclib/code/trunk@150 5acbf244-8a03-4a8b-a19b-0d601add4d27