Fixed IR intensity bug
It is possible for Gaussian to fail to compute the IR intensity, recording '******' in it's place in the log file.
This broke the parser as it assumed it would always find a string that could be converted to a float. This patch
means we return a 'nan' if the string is not a number.